backend/pyproject.toml
[project]
name = "jaldibharo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi[standard]>=0.135.2",
    "google-genai>=1.70.0",
    "google-generativeai>=0.8.6",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=5.0.0",
    "sqlmodel>=0.0.37",
    "tavily-python>=0.7.23",
    "uvicorn>=0.42.0",
]
